+++ R/dbscan.R
@@ -180,6 +180,7 @@
 #' res
 #'
 #' pairs(iris, col = res$cluster + 1L)
+#' clplot(iris, res)
 #'
 #' ## Use a precomputed frNN object
 #' fr <- frNN(iris, eps = .7)
@@ -200,6 +201,7 @@
 #' plot(x, col = res$cluster)
 #' points(x[res$cluster == 0, ], pch = 3, col = "grey")
 #'
+#' clplot(x, res)
 #' hullplot(x, res)
 #'
 #' ## Predict cluster membership for new data points