+++ R/hdbscan.R
@@ -144,12 +144,12 @@
 #' res
 #'
 #' plot(res)
-#' plot(moons, col = res$cluster + 1L)
+#' clplot(moons, res)
 #'
 #' ## cluster the moons data set with HDBSCAN using Manhattan distances
 #' res <- hdbscan(dist(moons, method = "manhattan"), minPts = 5)
 #' plot(res)
-#' plot(moons, col = res$cluster + 1L)
+#' clplot(moons, res)
 #'
 #' ## DS3 from Chameleon
 #' data("DS3")
@@ -159,9 +159,9 @@
 #'
 #' ## Plot the simplified tree, highlight the most stable clusters
 #' plot(res, show_flat = TRUE)
-#'
+#
 #' ## Plot the actual clusters (noise has cluster id 0 and is shown in black)
-#' plot(DS3, col = res$cluster + 1L, cex = .5)
+#' clplot(DS3, res)
 #'
 #' ## Example for HDBSCAN(e) using cluster_selection_epsilon
 #' # data with clusters of various densities.